feat(backend): Add validation and custom response

+++ src/core/HR.LeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
 using HR.LeaveManagement.Domain;
@@ -17,6 +18,9 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
         var leaveType = await _leaveTypeRepository.Get(request.DeleteLeaveTypeDto.Id);
+
+        if (leaveType == null)
+            throw new NotFoundException(nameof(LeaveType), request.DeleteLeaveTypeDto.Id);
         await _leaveTypeRepository.Delete(leaveType);
         return Unit.Value;
     }
